fix race condition in test/shutdown.js
It had assumed that a 5 publisher-1 subscriber topology would receive
messages in order (which on Travis CI seems to not occur) so had unsafe
shutdown code that would clean up prematurely in the case of late
deliverance.  This would mess up the count of subscribers since the
messages effectly race from publishers to subscriber.

As an example: before this patch, messages could be delivered in this
order:

p1 x10
p5
<shutdown occurs>
p2
<error, count is all messed up>

Now, each socket checks to see that the last one out hits the lights.