make PollCtx subclass Nan::AsyncWorker
This allows PollCtx to leave the lifetime management of the callback to
AsyncWorker, and allows us to use the AsyncWorker's async_resource.

No change in `make perf` was observed, though it's a little noisy.

We must implement the pure virtual function interface for Exectute(),
but I'm not sure what it's for.

Also, I believe that ~AsyncWorker() deletes the passed callback, as all
of the examples show calls to new Nan::Callback().

Fixes #209